refactor(am): split send outcome and peer routing seams
