+++ crates/atm-core/src/send/outcome.rs
@@ -1,0 +1,96 @@
+//! Public outcome types for the canonical mailbox write service.
+
+use serde::{Deserialize, Serialize};
+
+use crate::error_codes::AtmErrorCode;
+use crate::schema::AtmMessageId;
+use crate::types::{AgentName, CommandAction, TaskId, TeamName};
+
+fn is_false(value: &bool) -> bool {
+    !*value
+}
+
+/// Result of sending one ATM mailbox message.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct SendOutcome {
+    pub action: CommandAction,
+    pub team: TeamName,
+    pub agent: AgentName,
+    pub sender: AgentName,
+    pub outcome: SendCommandOutcome,
+    pub message_id: AtmMessageId,
+    pub requires_ack: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub task_id: Option<TaskId>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub summary: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub warnings: Vec<WarningEntry>,
+    #[serde(default, skip_serializing_if = "is_false")]
+    pub dry_run: bool,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "snake_case")]
+pub enum SendCommandOutcome {
+    Sent,
+    DryRun,
+}
+
+impl SendCommandOutcome {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            Self::Sent => "sent",
+            Self::DryRun => "dry_run",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct WarningEntry {
+    pub message: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub code: Option<AtmErrorCode>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub recovery: Option<String>,
+}
+
+impl WarningEntry {
+    pub fn new(message: impl Into<String>, recovery: Option<impl Into<String>>) -> Self {
+        Self {
+            message: message.into(),
+            code: None,
+            recovery: recovery.map(Into::into),
+        }
+    }
+
+    pub fn with_code(
+        code: AtmErrorCode,
+        message: impl Into<String>,
+        recovery: Option<impl Into<String>>,
+    ) -> Self {
+        Self {
+            message: message.into(),
+            code: Some(code),
+            recovery: recovery.map(Into::into),
+        }
+    }
+
+    pub fn render(&self) -> String {
+        let message = match self.code {
+            Some(code) if !self.message.contains(code.as_str()) => {
+                format!("{} [{}]", self.message, code.as_str())
+            }
+            _ => self.message.clone(),
+        };
+        match &self.recovery {
+            Some(recovery) if !message.contains("Recovery:") => {
+                format!("{message} Recovery: {recovery}")
+            }
+            None => message,
+            Some(_) => message,
+        }
+    }
+}
